Disables the user modules compilation
For the omni completion, the user will be in charge of compiling its own
project. This is much better if he uses Rebar or its project is complex
and uses special flags. Vimerl just need to find the .beam files,
usually reading rebar.config or look into $PWD/ or ebin/.