Read functions without specs from src correctly
So far we could only read functions with Edoc or specs from sources (and
could read the functions without these only from the beam). Now we can
read them from the source as well.

In practice, this means that if a source file is modified without being
recompiled, so far we could complete only the new functions with specs.
Now we can also compile new functions without specs.