Give proper error when "rebar path" fails
Previously when the "rebar path" command failed, the error message was
interpreted as a successful return value. The completion afterwards
usually failed (since the paths were not set).

Now when "rebar path" fails (i.e., its exit code is non-zero), its error
message is handled as if it belonged to the first line of the source
file. This causes the plugin to show the error properly to the user.