Make it possible to extend code path with configurable wildcard
This commit adds a configuration variable that makes it possible to
extend the code path that is used to search for modules. All the
directories that matches the wildcard in the variable will be added to
the code path. This option is useful when vim-erlang-omnicomplete cannot
figure out where modules are based on the project structure. For
example, the EMQX project (https://github.com/emqx/emqx) uses a fork of
rebar3 for building which vim-erlang-omnicomplete cannot handle.